direction moved to localStorage - new file for holding controls related code
